requirements.txt
Flask==3.0.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Flask-WTF==1.2.1
bcrypt==4.1.2
Flask-Limiter==3.5.0
limits==3.7.0
ordered-set==4.1.0 